# [RyuJIT] Unroll StartsWith/SequenceEqual for ([ReadOnly]Span<char>, const-string)