--- a/src/tests/JIT/opt/Unroll/Span_SequenceEqual_ConstString.csproj
+++ b/src/tests/JIT/opt/Unroll/Span_SequenceEqual_ConstString.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <Optimize>True</Optimize>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="Span_SequenceEqual_ConstString.cs" />
+  </ItemGroup>
+</Project>