… src/coreclr/jit/namedintrinsiclist.h
… src/libraries/System.Private.CoreLib/src/System/MemoryExtensions.Globalization.cs
… src/libraries/System.Private.CoreLib/src/System/MemoryExtensions.cs
… src/libraries/System.Private.CoreLib/src/System/String.cs
… src/tests/JIT/opt/Unroll/Span_SequenceEqual_ConstString.cs
… src/tests/JIT/opt/Unroll/Span_StartsWith_ConstString.cs
… src/tests/JIT/opt/Unroll/Span_StartsWith_ConstString.csproj