This PR optimizes the following APIs:
```
bool MemoryExtensions.StartsWith(ReadOnlySpan<T>, ReadOnlySpan<T>)
bool MemoryExtensions.StartsWith(ReadOnlySpan<T>, ReadOnlySpan<T>, StringComparison)
bool MemoryExtensions.SequenceEqual(ReadOnlySpan<T>, ReadOnlySpan<T>)
bool MemoryExtensions.Equals(ReadOnlySpan<T>, ReadOnlySpan<T>, StringComparison)
```
when the second arg is a constant string. For now it only handles strings with length of 1, 2 or 4 chars, but it can be extended to handle longer strings as well via SIMD (`[8..32]` range), here are some benchmarks.

It addresses https://github.com/dotnet/runtime/issues/45613 but only for spans since we promote them for high-performance tasks. In theory, it can be then extended to support plain string objects if needed.

#### 1. Small strings (e.g. length is 4)
```csharp
[Benchmark]
[Arguments("https://google.com")]
public bool SpanStartsWith(string str) => str.AsSpan().StartsWith("http");
```
```
       |       Mean |
       |-----------:|
master |  4.0456 ns |
    PR |  0.3910 ns |   10x faster
```
Codegen diff example: https://www.diffchecker.com/BeNpMF78

#### 2. Small strings (e.g. length is 4, ignore case)
```csharp
[Benchmark]
[Arguments("https://google.com")]
public bool SpanStartsWith(string str) => str.AsSpan().StartsWith("http", StringComparison.OrdinalIgnoreCase);
```
```
         |       Mean |
         |-----------:|
  master | 19.7386 ns |
      PR |  0.5509 ns |   35x faster
```
#### 3. Bigger strings (e.g. length is 23)
```csharp
[Benchmark]
public bool SpanStartsWith(string str) => str.AsSpan().StartsWith("ProxyAuthenticateHeader");
```
```
         |       Mean |
         |-----------:|
  master |  3.1628 ns |
proposed |  0.5293 ns |   5x faster (two AVX2 vectors)
```
#### 2. Bigger strings (e.g. length is 23, ignore case)
```csharp
[Benchmark]
public bool SpanStartsWith(string str) => 
    str.AsSpan().StartsWith("ProxyAuthenticateHeader", StringComparison.OrdinalIgnoreCase);
```
```
         |       Mean |
         |-----------:|
  master | 40.4105 ns |
proposed |  0.6116 ns |   66x faster (two AVX2 vectors)
```

Standalone benchmark: https://gist.github.com/EgorBo/8a4e4cda14eac0e605dd7bac68c56314

Inspired by LLVM: https://godbolt.org/z/8fcqfb

/cc @jkotas @dotnet/jit-contrib