@EgorBo 2020-12-25
For SIMD we only need to emit something like this in JIT:
```
*  ASG       simd32 (copy)
+--*  LCL_VAR   simd32<Vector256`1[UInt16]> V06 tmp2
\--*  HWINTRINSIC simd32 ushort Or
   +--*  HWINTRINSIC simd32 ushort Xor
   |  +--*  LCL_VAR   simd32<Vector256`1[UInt16]> V05 tmp1
   |  \--*  HWINTRINSIC simd32 ushort Create (.....)
   \--*  HWINTRINSIC simd32 ushort Xor
      +--*  LCL_VAR   simd32<Vector256`1[UInt16]> V03 loc0
      \--*  HWINTRINSIC simd32 ushort Create (.....)

*  RETURN    int
\--*  HWINTRINSIC bool   ushort TestZ
   +--*  LCL_VAR   simd32<Vector256`1[UInt16]> V06 tmp2
   \--*  LCL_VAR   simd32<Vector256`1[UInt16]> V06 tmp2
```