@EgorBo 2021-03-08
> @EgorBo My thought is that this is a lot of code for a very specific set of APIs, and quite restricted compared to the generality of the specified APIs (e.g., only certain types of strings). Is it sufficiently motivated? Namely, is there no other way for users to achieve better performance with a source-level implementation that would be "good enough", without the need to support all-platforms with JIT changes? Could the JIT do better (even if not optimal) with the general case without converting these to intrinsics? Could the source implementation use already existing hardware intrinsics? Could a new, specific, set of source-level APIs be defined to handle these special cases, not requiring JIT changes to achieve the perf results?

Ok, makes sense going to close it.

PS: Still, this benchmark:
```csharp
[Benchmark]
public bool SpanStartsWith(string str) => 
    str.AsSpan().StartsWith("ProxyAuthenticateHeader", StringComparison.OrdinalIgnoreCase);
```
```
         |       Mean |
         |-----------:|
  master | 40.4105 ns |
proposed |  0.6116 ns |   66x faster (two AVX2 vectors)
```
is a sign we can do better there.